LayoutTests/fast/events/ios/keyboard-event-key-attribute-expected.txt

Tests that KeyboardEvent.key is properly initialized on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.key is "a"
PASS event.code is "KeyA"
PASS event.key is "@"
PASS event.code is "Digit2"
PASS successfullyParsed is true

TEST COMPLETE
